fix: Fix re-shuffle in quiz when re-render.
