+++ src/app/dictionary/quiz/page.tsx
@@ -1,8 +1,15 @@
+'use client';
+
 import React from 'react';
 import Quiz from './quiz';
+import { useWords } from '../wordsContext';
 
 export default function QuizMain() {
+  const words = useWords();
+
   return (
-    <Quiz />
+    <div>
+      {words.length > 0 && <Quiz words={words} />}
+    </div>
   );
 }
